test(data/remotelistfavoritelocations): refactor test with makeSut and moving factory to your folder
